sentinel_ai/requirements.txt
# Sentinel AI - Production Dependencies

# Core dependencies
pyyaml>=6.0
psutil>=5.9.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0

# Deep Learning — Keras LSTM Autoencoder (time-series anomaly detection)
# Uses PyTorch as backend (TensorFlow not yet available for Python 3.14)
torch>=2.10.0
keras>=3.0.0

# MQTT
paho-mqtt>=1.6.1

# AWS Integration
boto3>=1.26.0
botocore>=1.29.0

# AI Integration (Ollama - local, free)
ollama>=0.3.0

# OpenAI Integration (optional, requires billing)
openai>=1.0.0

# Web Dashboard
Flask>=2.3.0
Flask-CORS>=4.0.0

# Optional but recommended
python-dateutil>=2.8.2

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
